requirements.txt
berserk==0.10.0
certifi==2022.9.24
charset-normalizer==2.1.1
chess==1.9.3
click==8.1.3
Deprecated==1.2.13
gTTS==2.2.4
idna==3.4
munch==2.5.0
ndjson==0.3.1
playsound==1.3.0
python-dotenv==0.21.0
requests==2.28.1
six==1.16.0
text-to-speech==1.5.3
urllib3==1.26.12
wrapt==1.14.1
